+++ src/main/java/duke/command/EndCommand.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.exception.DukeException;
+
 public class EndCommand extends Command {
 
     public static final String END = "Bye! Hope you had fun!";
