+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.component.TaskList;
+import duke.exception.DukeException;
+import duke.task.ToDo;
+
 public class ToDoCommand extends Command {
 
     public ToDoCommand(String content, TaskList tasks) {
