+++ src/main/java/duke/component/Parser.java
@@ -1,3 +1,9 @@
+package duke.component;
+
+import duke.command.*;
+import duke.exception.DukeException;
+
+
 public class Parser {
 
     public Command parse(String message, TaskList tasks) throws DukeException {
